requirements.txt
streamlit==1.54.0
fastapi
uvicorn
google-generativeai>=0.8.0
python-docx==1.2.0
requests>=2.31.0
python-dotenv>=1.0.1